chore: delete LoginInputGroup
